WIP moving business logic and db queries to models.
