--- app/Models/Keywords.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Keywords extends Model
-{
-    //
-}
