+++ app/News.php
@@ -1,17 +1,34 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App;
 
-use Illuminate\Http\Request;
+use Illuminate\Database\Eloquent\Model;
 use GuzzleHttp\Client;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 
-
-class HackerNewsController extends Controller
+class News extends Model
 {
-    private function CurateArticlesByTitle($articleList, $user_id) 
+    public function GetArticles()
     {
+        $client = new Client();
+        $topStoriesResponse = $client->get('https://hacker-news.firebaseio.com/v0/topstories.json');
+        $topStoryIDs = json_decode($topStoriesResponse->getBody(), true); 
+        $top25IDs = array_slice($topStoryIDs, 0, 25);
+
+        $top25 = [];
+
+        foreach ($top25IDs as $id) {
+            $articleResponse = $client->get("https://hacker-news.firebaseio.com/v0/item/$id.json");
+            $article = json_decode($articleResponse->getBody(), true);
+            array_push($top25, $article);
+        }
+        return $top25;
+    }
+
+    public function GetCuratedArticles($userID)
+    {
+        $top25 = $this->GetArticles();
         $curatedArticles = [];
         $keywords = DB::select('select keyword from keywords where user_id = ?', [$user_id]);
         $threshold = DB::select('select comment_threshold from user_comment_threshold where user_id = ?', [$user_id]);
@@ -32,30 +49,4 @@ class HackerNewsController extends Controller
         }
         return $curatedArticles;
     }
-
-    public function GetArticles(Request $req) 
-    {
-        $client = new Client();
-        $topStoriesResponse = $client->get('https://hacker-news.firebaseio.com/v0/topstories.json');
-        $topStoryIDs = json_decode($topStoriesResponse->getBody(), true); 
-        $top25IDs = array_slice($topStoryIDs, 0, 25);
-
-        $top25 = [];
-
-        foreach ($top25IDs as $id) {
-            $articleResponse = $client->get("https://hacker-news.firebaseio.com/v0/item/$id.json");
-            $article = json_decode($articleResponse->getBody(), true);
-            array_push($top25, $article);
-        }
-        
-        // If user is logged in, return articles based on curation criteria. Else, return top 25 from Hacker News
-        if ($req->userID) {
-            return $this->CurateArticlesByTitle($top25, $req->userID);     
-        } else {
-            return $top25;
-        }
-
-    } 
-
 }
-
